+++ R/zzz.R
@@ -27,6 +27,7 @@
 .onLoad = function(libname, pkgname) { # nolint
   # nocov start
   backports::import(pkgname)
+  backports::import(pkgname, "R_user_dir", force = TRUE)
   mlr3::mlr_tasks$add("oml", OMLTaskConnector)
   mlr3::mlr_resamplings$add("oml", OMLResamplingConnector)
 
@@ -35,9 +36,14 @@
   if (Sys.getenv("IN_PKGDOWN") == "true") {
     lg$set_threshold("warn")
   }
+
+  utils::globalVariables(c("is_target", "is_ignore", "is_row_identifier"), pkgname)
 } # nocov end
 
 .onUnload <- function (libpath) { # nolint
   # nocov start
   library.dynam.unload("mlr3oml", libpath)
 } # nocov end
+
+
+leanify_package()