+++ NEWS.md
@@ -1,6 +1,7 @@
 # mlr3oml 0.2.0
 
 * Added fallback to RWeka for sparse ARFF files (#6).
+* Fixed import from backports.
 
 # mlr3oml 0.1.0